router: introduce discovery_enable/disable
Discovery disabling may be useful when there are very many routers
and buckets, and a user does not want to pay overhead of the
automatic massive discovery. It may be expensive in big clusters.

In that case users may want to turn off discovery when there is
no rebalancing, and turn it back on, when it starts, to keep the
routers up to date with _bucket changes.

Part of #210

@TarantoolBot document
Title: vshard.router.discovery_enable/disable, and new cfg option

```Lua
vshard.router.discovery_disable()
```
Turns off the background discovery fiber used by the router to
find buckets. If discovery is already disabled, nothing happens.

```Lua
vshard.router.discovery_enable()
```
Turns on the background discovery fiber. If it is already turned
on, nothing happens.

If `vshard.router.discovery_wakeup()` is called when it is
disabled, nothing happens.

These methods are good to enable/disable discovery after the
router is already started, but it is enabled by default. You may
want to never enable it even for a short time - then specify
`discovery_enable` option in the configuration. Set to `true` to
enable discovery, and to `false` to disable. When it is disabled
through the option, discovery is never even started in
`vshard.router.cfg()`.

You may decide to turn off discovery if you have many routers, or
tons of buckets (hundreds of thousands and more), and you see that
the discovery process consumes notable CPU % on routers and
storages. In that case it may be wise to turn off discovery when
there is no rebalancing in the cluster. And turn it on for new
routers, as well as for all routers when rebalancing is started.